+++ common/game/setup-game.ts
